+++ src/tinycore/utilities/string.h
@@ -42,9 +42,9 @@ public:
     static StringVector split(const std::string &s, const std::string &delim, bool keepEmpty=true);
     static StringVector splitLines(const std::string &s, bool keepends=false);
 
-    static int count(const std::string &s, char c);
-    static int count(const std::string &s, const char *sub);
-    static int count(const std::string &s, std::function<bool (char)> pred);
+    static size_t count(const std::string &s, char c);
+    static size_t count(const std::string &s, const char *sub);
+    static size_t count(const std::string &s, std::function<bool (char)> pred);
 
     static void capitalize(std::string &s);
     static std::string capitalizeCopy(const std::string &s);
